# Extend Encryption API, Improve KeyGen performance