--- a/paillier.go
+++ b/paillier.go
@@ -86,11 +86,31 @@ func l(u *big.Int, n *big.Int) *big.Int {
 // Encrypt encrypts a plain text represented as a byte array. The passed plain
 // text MUST NOT be larger than the modulus of the passed public key.
 func Encrypt(pubKey *PublicKey, plainText []byte) ([]byte, error) {
+	c, _, err := EncryptAndNonce(pubKey, plainText)
+	return c, err
+}
+
+// EncryptAndNonce encrypts a plain text represented as a byte array, and in
+// addition, returns the nonce used during encryption. The passed plain text
+// MUST NOT be larger than the modulus of the passed public key.
+func EncryptAndNonce(pubKey *PublicKey, plainText []byte) ([]byte, *big.Int, error) {
 	r, err := rand.Int(rand.Reader, pubKey.N)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
+	c, err := EncryptWithNonce(pubKey, r, plainText)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return c.Bytes(), r, nil
+}
+
+// EncryptWithNonce encrypts a plain text represented as a byte array using the
+// provided nonce to perform encryption. The passed plain text MUST NOT be
+// larger than the modulus of the passed public key.
+func EncryptWithNonce(pubKey *PublicKey, r *big.Int, plainText []byte) (*big.Int, error) {
 	m := new(big.Int).SetBytes(plainText)
 	if pubKey.N.Cmp(m) < 1 { // N < m
 		return nil, ErrMessageTooLong
@@ -106,7 +126,7 @@ func Encrypt(pubKey *PublicKey, plainText []byte) ([]byte, error) {
 		pubKey.NSquared,
 	)
 
-	return c.Bytes(), nil
+	return c, nil
 }
 
 // Decrypt decrypts the passed cipher text.