This PR extends the encryption API of this package with:
 - `EncryptWithNonce` - Encrypts a plaintext with user provided nonce.
 - `EncryptAndNonce` - Generates a random nonce, and returns it to the user along with the ciphertext.

It also makes a simple optimization to generate `p` and `q` in parallel during keygen, offering a  modest speed up.

Primes generate concurrently:
```
goos: darwin                                                       
goarch: amd64                                                      
pkg: github.com/roasbeef/go-go-gadget-paillier                     
BenchmarkKey1024-8           100          16601491 ns/op          878931 B/op       3873 allocs/op                                     
BenchmarkKey2048-8            10         126341031 ns/op         1913843 B/op       6138 allocs/op                                     
BenchmarkKey3072-8             3         615750569 ns/op         4517426 B/op      11692 allocs/op                                     
BenchmarkKey4096-8             2        2607939290 ns/op        11143432 B/op      23340 allocs/op                                     
PASS                                                               
ok      github.com/roasbeef/go-go-gadget-paillier       12.398s 
```

Primes generated serially:
```
goos: darwin
goarch: amd64
pkg: github.com/roasbeef/go-go-gadget-paillier
BenchmarkKey1024-8            50          26715935 ns/op          972709 B/op       4209 allocs/op
BenchmarkKey2048-8            10         168290016 ns/op         1655725 B/op       5373 allocs/op
BenchmarkKey3072-8             3         803726192 ns/op         4932906 B/op      12703 allocs/op
BenchmarkKey4096-8             1        3289647242 ns/op        10710200 B/op      22465 allocs/op
PASS
ok      github.com/roasbeef/go-go-gadget-paillier       9.172s
```

Finally, the benchmarks have been modified to report allocations and properly reset the timer after initializing the relevant benchmarks. Several calls to `Fatalf` have also been corrected to use a formatting directive, allowing the package and tests to used with Go1.11. Without the directive, the compilation would fail due to the stricter vet checks at compile time.